backend/requirements.txt
fastapi
python-multipart
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]